requirements.txt
pandas>=2.0.0
requests>=2.28.0
stanza>=1.5.0
tqdm>=4.65.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
